refactor: move shared logic into shared package
