+++ internal/backup_restore/README.md
@@ -9,4 +9,4 @@ and into this repository. This file isn't intended to be modified.
 If `meta.pb.go` ever needs to be re-generated, follow these steps:
 1. Install `protoc` (i.e. via `brew install protobuf`)
 2. Run `go install github.com/gogo/protobuf/protoc-gen-gogo` from within this repository
-3. Run `go generate <path to clients/backup>`
+3. Run `go generate <path to internal/backup_restore>`
